Evolve particles and hide UI
Particles can now "evolve", as in change type to another random type. The ID/seed for the type to evolve into is set during the generation of each type. So all particles of the same type evolve into the same type.
If not moving the mouse or clicking any buttons for 10 seconds, the tools-UI will move off screen.